+++ src/constants.ts
@@ -1,0 +1,7 @@
+export const APP_NAME = "gcp-switch-config";
+export const GCP_SWITCH_COMMAND = `${APP_NAME}.open-gcp-switch-config`;
+
+export const APPLICATION_DEFAULT_CREDENTIALS_FILE_PATH =
+  ".config/gcloud/application_default_credentials.json";
+
+export const CACHE_VERSION = `${APP_NAME}_cache_v1`;